+++ pueue/src/bin/pueued.rs
@@ -5,8 +5,7 @@ use clap::Parser;
 use log::warn;
 use simplelog::{Config, ConfigBuilder, LevelFilter, SimpleLogger, TermLogger, TerminalMode};
 
-use pueue::daemon::cli::CliArguments;
-use pueue::daemon::run;
+use pueue::daemon::{cli::CliArguments, run};
 
 #[tokio::main(flavor = "multi_thread", worker_threads = 4)]
 async fn main() -> Result<()> {
@@ -14,6 +13,15 @@ async fn main() -> Result<()> {
     let opt = CliArguments::parse();
 
     if opt.daemonize {
+        // Ordinarily this would be handled in clap, but they don't support conflicting specific args
+        // with subcommands. We can't turn this off globally because -c and -p are valid args when using
+        // subcommand to install the service
+        #[cfg(target_os = "windows")]
+        if opt.service.is_some() {
+            println!("daemonize flag cannot be used with service subcommand");
+            return Ok(());
+        }
+
         return fork_daemon(&opt);
     }
 
@@ -47,6 +55,44 @@ async fn main() -> Result<()> {
         SimpleLogger::init(level, logger_config).unwrap();
     }
 
+    #[cfg(target_os = "windows")]
+    {
+        use pueue::daemon::cli::{ServiceSubcommand, ServiceSubcommandEntry};
+        use pueue::daemon::service;
+
+        if let Some(ServiceSubcommandEntry::Service(service)) = opt.service {
+            match service {
+                ServiceSubcommand::Run => {
+                    // start service
+                    service::run_service(opt.config.clone(), opt.profile.clone())?;
+                    return Ok(());
+                }
+
+                ServiceSubcommand::Install => {
+                    service::install_service(opt.config.clone(), opt.profile.clone())?;
+                    println!("Successfully installed `pueued` Windows service");
+                    return Ok(());
+                }
+
+                ServiceSubcommand::Uninstall => {
+                    service::uninstall_service()?;
+                    println!("Successfully uninstalled `pueued` Windows service");
+                    return Ok(());
+                }
+
+                ServiceSubcommand::Start => {
+                    service::start_service()?;
+                    return Ok(());
+                }
+
+                ServiceSubcommand::Stop => {
+                    service::stop_service()?;
+                    return Ok(());
+                }
+            }
+        }
+    }
+
     run(opt.config, opt.profile, false).await
 }
 
@@ -78,7 +124,21 @@ fn fork_daemon(opt: &CliArguments) -> Result<()> {
         "pueued".to_string()
     };
 
-    Command::new(current_exe)
+    let mut command = Command::new(current_exe);
+
+    #[cfg(target_os = "windows")]
+    {
+        use std::os::windows::process::CommandExt;
+        const CREATE_NO_WINDOW: u32 = 0x08000000;
+        // CREATE_NO_WINDOW causes all children to not show a visible console window,
+        // but it also apparently has the effect of starting a new process group.
+        //
+        // https://learn.microsoft.com/en-us/windows/win32/procthread/process-creation-flags#flags
+        // https://stackoverflow.com/a/71364777/9423933
+        command.creation_flags(CREATE_NO_WINDOW);
+    }
+
+    command
         .args(&arguments)
         .spawn()
         .expect("Failed to fork new process.");